Fix test_huggingface_client
